backend/requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
marshmallow
marshmallow-sqlalchemy
PyJWT
Werkzeug
dnspython
